Module move_prover::cli

Functionality related to the command line interface of the move prover.

Structs

Options
Represents options provided to the tool.

PreludeTemplateContext
Options used for prelude template. See prelude.bpl for documentation.

Enums

VerificationScope
Default for what functions to verify.

Constants

INLINE_PRELUDE
Represents the virtual path to the boogie prelude which is inlined into the binary.