Struct move_prover::cli::Options

pub struct Options {
    pub prelude_path: String,
    pub output_path: String,
    pub docgen: bool,
    pub docgen_options: DocgenOptions,
    pub account_address: String,
    pub verbosity_level: LevelFilter,
    pub move_sources: Vec<String>,
    pub move_deps: Vec<String>,
    pub search_path: Vec<String>,
    pub boogie_exe: String,
    pub z3_exe: String,
    pub use_cvc4: bool,
    pub cvc4_exe: String,
    pub boogie_flags: Vec<String>,
    pub generate_only: bool,
    pub native_stubs: bool,
    pub minimize_execution_trace: bool,
    pub omit_model_debug: bool,
    pub use_array_theory: bool,
    pub stable_test_output: bool,
    pub verify_scope: VerificationScope,
    pub template_context: PreludeTemplateContext,
    pub bench_repeat: usize,
    pub generate_smt: bool,
}

Represents options provided to the tool.

Fields
prelude_path: String

Path to the boogie prelude. The special string INLINE_PRELUDE is used to refer to a prelude build into this binary.

output_path: String

The path to the boogie output which represents the verification problem.

docgen: bool

Whether to run the documentation generator instead of the prover.

docgen_options: DocgenOptions

Options for the document generator.

account_address: String

An account address to use if none is specified in the source.

verbosity_level: LevelFilter

Verbosity level for logging.

move_sources: Vec<String>

The paths to the move sources.

move_deps: Vec<String>

The paths to any dependencies for the move sources. Those will not be verified but can be used by move_sources.

search_path: Vec<String>

Paths to directories where dependencies are looked up automatically.

boogie_exe: String

Path to the boogie executable.

z3_exe: String

Path to the z3 executable.

use_cvc4: bool

Whether to use cvc4.

cvc4_exe: String

Path to the cvc4 executable.

boogie_flags: Vec<String>

List of flags to pass on to boogie.

generate_only: bool

Whether to only generate boogie.

native_stubs: bool

Whether to generate stubs for native functions.

minimize_execution_trace: bool

Whether to minimize execution traces in errors.

omit_model_debug: bool

Whether to omit debug information in generated model.

use_array_theory: bool

Whether to use native array theory.

stable_test_output: bool

Whether output for e.g. diagnosis shall be stable/redacted so it can be used in test output.

verify_scope: VerificationScope

Scope of what functions to verify

template_context: PreludeTemplateContext

Template context for prelude. This is map from variable names into strings (or values represented as strings)

bench_repeat: usize

How many times to call boogie on the verification problem. This is used for benchmarking.

generate_smt: bool

Whether to produce an SMT file for each verification problem.

Methods

impl Options

pub fn initialize_from_args(&mut self, args: &[String]) -> Result<()>

pub fn setup_logging(&self)

Sets up logging based on provided options. This should be called as early as possible and before any use of info!, warn! etc.

pub fn setup_logging_for_test(&self)

pub fn get_boogie_command(&self, boogie_file: &str) -> Vec<String>

Returns command line to call boogie.

pub fn get_boogie_log_file(&self, boogie_file: &str) -> String

Returns name of file where to log boogie output.